+++ src/components/ItaliaTheme/Blocks/ArgumentsInEvidence/BottomBody.jsx
@@ -5,6 +5,7 @@ import {
   Chip,
   ChipLabel,
   Button,
+  Container,
 } from 'design-react-kit/dist/design-react-kit';
 
 const messages = defineMessages({
@@ -20,38 +21,37 @@ const messages = defineMessages({
 
 const BottomBody = ({ data, intl }) => {
   return data?.arguments?.length > 0 ? (
-    <>
-      <div className="col-lg-10 col-xl-8 offset-lg-1 offset-xl-3 pt-5">
-        <div className="row d-lg-inline-flex">
-          <div className="col-lg-3">
-            <h6 className="text-uppercase text-center mt-1">
-              {intl?.formatMessage(messages.otherArguments)}
-            </h6>
-          </div>
-          <div className="col-lg-9">
-            {data?.arguments?.map((argument, index) => (
-              <ConditionalLink
-                condition={!!argument['@id']}
-                item={argument}
-                key={index}
-                title={argument.title}
-                className="text-decoration-none"
+    <Container className="text-center">
+      <div className="row d-lg-inline-flex pt-5">
+        <div className="col-lg-3">
+          <h6 className="text-uppercase text-center mt-1">
+            {intl?.formatMessage(messages.otherArguments)}
+          </h6>
+        </div>
+        <div className="col-lg-9 text-left">
+          {data?.arguments?.map((argument, index) => (
+            <ConditionalLink
+              condition={!!argument['@id']}
+              item={argument}
+              key={index}
+              title={argument.title}
+              className="text-decoration-none"
+            >
+              <Chip
+                color="primary"
+                disabled={false}
+                large
+                simple
+                tag="div"
+                className="mr-2"
               >
-                <Chip
-                  color="primary"
-                  disabled={false}
-                  large
-                  simple
-                  tag="div"
-                  className="mr-2"
-                >
-                  <ChipLabel tag="span">{argument.title}</ChipLabel>
-                </Chip>
-              </ConditionalLink>
-            ))}
-          </div>
+                <ChipLabel tag="span">{argument.title}</ChipLabel>
+              </Chip>
+            </ConditionalLink>
+          ))}
         </div>
       </div>
+
       <div className="link-button mt-5">
         <UniversalLink href="/argomenti" className="text-decoration-none">
           <Button
@@ -64,7 +64,7 @@ const BottomBody = ({ data, intl }) => {
           </Button>
         </UniversalLink>
       </div>
-    </>
+    </Container>
   ) : null;
 };
 export default BottomBody;